Refactored exception namespaces - moved session exceptions in Bead\Exceptions\Session namespace - removed database exceptions that had already been copied to their own namespace
