--- src/Exceptions/QueryBuilderException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Bead\Exceptions;
-
-use Exception;
-
-class QueryBuilderException extends Exception
-{}
