+++ src/Exceptions/Session/InvalidSessionDirectoryException.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace Bead\Exceptions;
+namespace Bead\Exceptions\Session;
 
-use Exception;
 use Throwable;
 
 /**
